Narrow down the architectures to where the might compile
The upstream has a short list which is supported.
Removing all architectures which is explicitly unsupported creates less
noises on the build hosts. There are still architectures, which is
theoretically supported, but the tests fails. But I'm keeping that
architectures in the list now and I'm going to try fixing them.